pr-tool: move github functionality into separate module
